+++ rapidsms_xforms/urls.py
@@ -4,12 +4,12 @@ from . import views
 
 urlpatterns = patterns('',
     url(r"^xforms/$", login_required(views.xforms), name="xforms"),
-    url(r"^xforms/new/$", login_required(views.new_xform)),
+    url(r"^xforms/new/$", login_required(views.new_xform), name='xforms_create'),
     url(r"^xforms/(\d+)/submissions/$", login_required(views.view_submissions)),
     url(r"^xforms/submissions/(\d+)/edit/$", login_required(views.edit_submission)),
     url(r"^xforms/(\d+)/view/$", login_required(views.view_form)),
     url(r"^xforms/(\d+)/details/$", login_required(views.view_form_details)),
-    url(r"^xforms/(\d+)/edit/$", login_required(views.edit_form)),
+    url(r"^xforms/(\d+)/edit/$", login_required(views.edit_form), name='xforms_edit'),
     url(r"^xforms/(\d+)/order/$", login_required(views.order_xform)),
     url(r"^xforms/(\d+)/delete/$", login_required(views.delete_xform)),
     url(r"^xforms/(\d+)/add_field/$", login_required(views.add_field)),
@@ -25,9 +25,9 @@ urlpatterns = patterns('',
     url(r"^xforms/(\d+)/delete_field/(\d+)/$", login_required(views.delete_field)),
      
     # these are ODK URLs to be used by ODK Collect
-    url(r"^formList$", views.odk_list_forms),
-    url(r"^xforms/odk/get/(\d+)/$", views.odk_get_form),
-    url(r"^submission", views.odk_submission),
+    url(r"^formList$", views.odk_list_forms, name='odk_list'),
+    url(r"^xforms/odk/get/(\d+)/$", views.odk_get_form, name='odk_form'),
+    url(r"^submission", views.odk_submission, name='odk_submit'),
 
     # CSV Export
     url(r"^xforms/(\d+)/submissions.csv$", views.submissions_as_csv)